+++ src/app/cart/services/cart.service.ts
@@ -8,46 +8,64 @@ import { ProductCartModel } from '../../products/models/products.model';
 })
 export class CartService {
   public itemsChanged$: Subject<any> = new Subject<any>();
-  private purchasedItems: Array<ProductCartModel> = [];
+  private cartProducts: Array<ProductCartModel> = [];
+  private totalQuantity = 0;
+  private totalSum = 0;
 
   constructor() { }
 
   getProducts(): ProductCartModel[] {
-    return [...this.purchasedItems];
+    return [...this.cartProducts];
   }
 
   addProduct(item: ProductCartModel): void {
-    const exist = this.purchasedItems.findIndex(i => i.id === item.id);
+    const exist = this.cartProducts.findIndex(i => i.id === item.id);
     if (exist !== -1) {
-      this.purchasedItems[exist].quantity += 1;
+      this.cartProducts[exist].quantity += 1;
       this.calcTotalPrice(exist);
     } else {
       item.quantity = 1;
       item.totalPrice = item.price;
-      this.purchasedItems.push(item);
+      this.cartProducts.push(item);
     }
+    this.updateCartData();
     this.itemsChanged$.next();
   }
 
-  deleteProduct(id: number): void {
-    this.purchasedItems = this.purchasedItems.filter(i => i.id !== id);
+  removeProduct(id: number): void {
+    this.cartProducts = this.cartProducts.filter(i => i.id !== id);
+    this.updateCartData();
     this.itemsChanged$.next();
   }
 
-  updateProduct(id: number, key: string, value: any): void {
-    const exist = this.purchasedItems.findIndex(i => i.id === id);
+  changeQuantity(id: number, value: any): void {
+    const exist = this.cartProducts.findIndex(i => i.id === id);
     if (exist !== -1) {
-      this.purchasedItems[exist][key] = value;
+      this.cartProducts[exist].quantity = value;
       this.calcTotalPrice(exist);
+      this.updateCartData();
       this.itemsChanged$.next();
     }
   }
 
-  getSummaryBy(key: string): number {
-    return this.purchasedItems.reduce((prev, curr) => prev + curr[key], 0);
+  getTotalPrice(): number {
+    return this.totalSum;
+  }
+
+  getTotalQuantity(): number {
+    return this.totalQuantity;
+  }
+
+  isEmptyCart(): boolean {
+    return this.cartProducts.length === 0;
   }
 
   private calcTotalPrice(index: number): void {
-    this.purchasedItems[index].totalPrice = this.purchasedItems[index].quantity * this.purchasedItems[index].price;
+    this.cartProducts[index].totalPrice = this.cartProducts[index].quantity * this.cartProducts[index].price;
+  }
+
+  private updateCartData(): void {
+    this.totalQuantity = this.cartProducts.reduce((prev, curr) => prev + curr.quantity, 0);
+    this.totalSum = this.cartProducts.reduce((prev, curr) => prev + curr.totalPrice, 0);
   }
 }