+++ src/app/shared/directives/styled-by-click.directive.spec.ts
@@ -1,0 +1,8 @@
+import { StyledByClickDirective } from './styled-by-click.directive';
+
+describe('StyledByClickDirective', () => {
+  it('should create an instance', () => {
+    const directive = new StyledByClickDirective();
+    expect(directive).toBeTruthy();
+  });
+});