+++ src/app/shared/shared.module.ts
@@ -2,12 +2,13 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
-import { HighlightDirective } from './directives/highlight.directive';
+
+import { HighlightDirective, StyledByClickDirective } from './directives';
 
 
 
 @NgModule({
-  declarations: [HighlightDirective],
+  declarations: [HighlightDirective, StyledByClickDirective],
   imports: [
     CommonModule,
     MatButtonModule,
@@ -17,6 +18,7 @@ import { HighlightDirective } from './directives/highlight.directive';
     MatButtonModule,
     MatCardModule,
     HighlightDirective,
+    StyledByClickDirective,
   ],
 })
 export class SharedModule { }